backend/pyproject.toml
[project]
name = "backend"
version = "1.0.1"
description = "GPU Connect - Decentralized AI Compute Marketplace. Connect providers and consumers to share GPU resources via secure token-based authentication and real-time WebSocket communication."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.13.3",
    "celery>=5.6.2",
    "channels>=4.3.2",
    "channels-redis>=4.3.0",
    "daphne>=4.2.1",
    "dj-database-url>=3.1.0",
    "django>=6.0.2",
    "django-allauth[socialaccount]>=65.5.0",
    "django-cors-headers>=4.9.0",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "gunicorn>=25.1.0",
    "psycopg[binary]>=3.3.2",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "pytest-django>=4.11.1",
    "python-dotenv>=1.0.1",
    "redis>=7.1.1",
    "requests>=2.32.5",
    "websockets>=16.0",
    "whitenoise>=6.11.0",
]

# ── Pylint Configuration ──────────────────────────────────────────────
[tool.pylint.main]
load-plugins = ["pylint_django"]
django-settings-module = "config.settings"
ignore-patterns = ["0\\d+.*\\.py"]     # ignore auto-generated migration files
ignore = ["manage.py", "custom_migrations"]

[tool.pylint."messages control"]
disable = [
    "import-error",                  # handled by Django / runtime imports
    "too-few-public-methods",        # DRF Meta classes, small test classes
    "attribute-defined-outside-init", # pytest setup_method pattern
    "duplicate-code",                # test setup boilerplate
    "broad-exception-caught",        # legitimate in WebSocket error handling
    "arguments-renamed",             # WebSocket method overrides
    "arguments-differ",              # WebSocket method overrides
    "import-outside-toplevel",       # lazy imports in consumers / manage.py
    "missing-function-docstring",    # test helper methods
]

[tool.pylint.format]
max-line-length = 120
